+++ tests/providers/apache/hive/transfers/test_hive_to_mysql.py
@@ -49,7 +49,9 @@ class TestHiveToMySqlTransfer(TestHiveEnvironment):
 
         mock_hive_hook.assert_called_once_with(hiveserver2_conn_id=self.kwargs["hiveserver2_conn_id"])
         mock_hive_hook.return_value.get_records.assert_called_once_with("sql", parameters={})
-        mock_mysql_hook.assert_called_once_with(mysql_conn_id=self.kwargs["mysql_conn_id"])
+        mock_mysql_hook.assert_called_once_with(
+            mysql_conn_id=self.kwargs["mysql_conn_id"], local_infile=False
+        )
         mock_mysql_hook.return_value.insert_rows.assert_called_once_with(
             table=self.kwargs["mysql_table"], rows=mock_hive_hook.return_value.get_records.return_value
         )
@@ -84,6 +86,7 @@ class TestHiveToMySqlTransfer(TestHiveEnvironment):
 
         HiveToMySqlOperator(**self.kwargs).execute(context=context)
 
+        mock_mysql_hook.assert_called_once_with(mysql_conn_id=self.kwargs["mysql_conn_id"], local_infile=True)
         mock_tmp_file_context.assert_called_once_with()
         mock_hive_hook.return_value.to_csv.assert_called_once_with(
             self.kwargs["sql"],