+++ airflow/providers/mysql/CHANGELOG.rst
@@ -27,6 +27,15 @@ used with MySQL server 5.6.4 through 5.7.
 Changelog
 ---------
 
+4.0.0
+.....
+
+Breaking Changes
+~~~~~~~~~~~~~~~~
+
+You can no longer pass "local_infile" as extra in the connection. You should pass it instead as
+hook's "local_infile" parameter when you create the MySqlHook (either directly or via hook_params).
+
 3.4.0
 .....